requirements.txt
PyQt6>=6.6
pyqtgraph>=0.13
numpy>=1.24
scipy>=1.11
sounddevice>=0.4
soundfile>=0.12
onnxruntime-directml>=1.17
librosa>=0.10
mido>=1.3
